rename: editor 폴더 변경
