+++ frontend/src/components/@commons/markdownEditor/UiViewer.stories.tsx
@@ -1,5 +1,5 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
-import UiViewer from 'components/@commons/UiViewer';
+import UiViewer from 'components/@commons/markdownEditor/UiViewer';
 
 export default {
   title: 'UiViewer',
