+++ frontend/src/components/preQuestion/PreQuestionViewModal.tsx
@@ -7,7 +7,7 @@ import closeIcon from 'assets/images/close.svg';
 import Button from 'components/@commons/Button';
 import FlexBox from 'components/@commons/FlexBox';
 import Image from 'components/@commons/Image';
-import UiViewer from 'components/@commons/UiViewer';
+import UiViewer from 'components/@commons/markdownEditor/UiViewer';
 import ModalPortal from 'portal/ModalPortal';
 import { PreQuestionCustomHookType, PreQuestionFormatType } from 'types/preQuestion';
 import { ParticipantType } from 'types/team';
